+++ wdmapper/format/csv.py
@@ -98,6 +98,7 @@ class Writer(LinkWriter):
     def __init__(self, stream, header=True):
         self.stream = stream
         self.header = header
+        self.meta = {}
         self.initialized = header is False
 
     def init(self, meta):