+++ wdmapper/property.py
@@ -41,8 +41,7 @@ class Property(object):
         self.template = data.get('template')
 
         # type could also be: string, commons-media, globe-coordinate
-        if (self.type != 'http://wikiba.se/ontology#ExternalId' and
-            self.type != 'http://wikiba.se/ontology#String'):
+        if (self.type not in ['http://wikiba.se/ontology#ExternalId', 'http://wikiba.se/ontology#String']):
             error = 'property {id} is not of type external-id or string!'
         elif not self.template:
             error = 'property {id} lacks URL template (P1630)'